requirements.txt
streamlit
pandas
numpy
plotly
folium
psycopg2-binary
sqlalchemy
xarray
netCDF4
sentence-transformers
faiss-cpu
langchain
python-dotenv
bcrypt
PyJWT
logging